Add transition slop zones and update tests
As the quadrants now have slop they are mutually exclusive from the axes. But some lines travel through an axis which should *need to* consider the axis.

This is particularly the case for graphs such as x^2 which would leave a quadrant and enter and axis before entering the origin. This is counter-intuitive.

These transition zones allow for the matching algorithm to catch these cases. By only adding in these zones in specific cases (that is where a line would need to transition from/into a zone from a quadrant) we can control whether to add or ignore axis leeway.