Remove transition zones from curve intersection
Curve intersection used the same code from the sectorBuilder as the axis intersection and sector inclusion. This meant that the transition zones were being added when no transition between zones is occurring, just single point classification.

The classification of points now ignores slop as a point cannot transition between zones.

The custom settings object would deserialise the setting strings through the sectorBuilder and thus add the transition zones which is not desired. (A point to note that in future we may want to add custom transition zones or dynamically create them, but few to no people use this code except us). This was excerbated in the checker tuner which would add the slop and print that to the custom settings box, thus each refresh just kept adding more and more slop ad infinitum.